apps/micro-services/cart/pyproject.toml
[project]
name = "cart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "flask-login>=0.6.2",
    "flask-wtf>=1.2.0",
    "flask-migrate>=4.0.0",
    "flask-mail>=0.9.1",
    "wtforms>=3.1.0",
    "werkzeug>=3.0.0",
    "python-dotenv>=1.0.0",
    "stripe>=7.0.0",
    "pillow>=10.0.0",
    "sqlalchemy>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "faker>=19.0.0",
]

